move coderenderer to it's own file
