+++ docs/configuration/sidebar.md
@@ -55,7 +55,6 @@ const highlight = (code, language) => {
 }
 ```
 
-
 You can infinitely nest folders, although we recommend limiting this to 2 or 3 levels otherwise navigating can become tedious and it's more difficult to visualize the hierarchy.
 
 ![alt text](/git-branching.png)
